move email and password validator widget in shared folder
